// src/types/todo.ts
//Типы для todoReducer

export interface TodoState {
    todos: any[];
    loading: boolean;
    error: null | string;
    page: number;//Текущая страница
    limit: number;//Всего страниц
}

//Перечисления для типов (для свитчей)
export enum TodoActionTypes {
    FETCH_TODOS = 'FETCH_TODOS',
    FETCH_TODOS_SUCCESS = 'FETCH_TODOS_SUCCESS',
    FETCH_TODOS_ERROR = 'FETCH_TODOS_ERROR',
    SET_TODO_PAGE = 'SET_TODO_PAGE'//Action для изменения текущей страницы
}

//Интерфейсы для каждого Action
interface FetchTodoAction { 
    type: TodoActionTypes.FETCH_TODOS;
}

interface FetchTodoSuccessAction { 
    type: TodoActionTypes.FETCH_TODOS_SUCCESS;
    payload: any[];
}

interface FetchTodoErrorAction { 
    type: TodoActionTypes.FETCH_TODOS_ERROR;
    payload: string;
}

interface SetTodoPage { 
    type: TodoActionTypes.SET_TODO_PAGE;
    payload: number;
}

//Объединение action воедино, универсальная типизация action
export type TodoAction = FetchTodoAction | FetchTodoSuccessAction | FetchTodoErrorAction | SetTodoPage;